feat(llm): add debug logging of full OpenAI request
Log the complete request body (model, messages, tools, etc.) at debug level
before sending to OpenAI API for easier debugging.
